python/requirements.txt
avro_python3==1.8.2
kafka-python==1.4.7
requests==2.22.0
